Finish supporting basics for Immutable checking
This commit fleshes out the ImmutableProcessor with logic for the basic
requirements for immutability checking according to the specs given in
Effective Java 3rd edition (item 17). Specifically, this annotation
processor adheres very strictly to the rules given. It's possible that
the checks could be relaxed a bit in future updates.

New in this change:
- Ensuring that fields are private and final.
- Ensuring that fields are themselves either primitive or @Immutable.
- Whitelisting java.lang.String fields so that they don't have to be
  annotated.
- Updating the @Immutable annotation type to allow it to be used on
  fields (necessary for the checking specified above).